hf-image-captioning/requirements.txt
streamlit
Pillow
# For transformers and its core backend:
transformers
torch
